+++ circus/commands/listsockets.py
@@ -1,4 +1,5 @@
 from circus.commands.base import Command
+import operator
 
 
 class ListSockets(Command):
@@ -17,7 +18,8 @@ class ListSockets(Command):
             }
 
 
-        The response return the list asked.
+        The response return a list of json mappings with keys for fd, name,
+        host and port.
 
         Command line
         ------------
@@ -32,12 +34,17 @@ class ListSockets(Command):
         return self.make_message()
 
     def execute(self, arbiter, props):
-        sockets = [str(socket) for socket in arbiter.sockets.values()]
-        sockets.sort()
+        sockets = [{'fd': socket.fileno(), 'name': socket.name,
+                    'host': socket.host, 'port': socket.port,
+                    'backlog': socket.backlog}
+                   for socket in arbiter.sockets.values()]
+        sockets.sort(key=operator.itemgetter('fd'))
         return {"sockets": sockets}
 
     def console_msg(self, msg):
         if 'sockets' in msg:
-            return "\n".join(msg['sockets'])
+            return "\n".join(['%d:socket %r at %s:%d' % (
+                s['fd'], s['name'], s['host'], s['port'])
+                for s in msg['sockets']])
 
         return self.console_error(msg)