+++ circus/stats/streamer.py
@@ -4,12 +4,13 @@ import json
 from itertools import chain
 import os
 import errno
+import socket
 
 from zmq.eventloop import ioloop, zmqstream
 
 from circus.commands import get_commands
 from circus.client import CircusClient
-from circus.stats.collector import StatsCollector
+from circus.stats.collector import WatcherStatsCollector, SocketStatsCollector
 from circus.stats.publisher import StatsPublisher
 from circus import logger
 
@@ -23,33 +24,24 @@ class StatsStreamer(object):
         self.sub_socket = self.ctx.socket(zmq.SUB)
         self.sub_socket.setsockopt(zmq.SUBSCRIBE, self.topic)
         self.sub_socket.connect(self.pubsub_endpoint)
-        self.loop = ioloop.IOLoop()  # events coming from circusd
+        self.loop = ioloop.IOLoop.instance()  # events coming from circusd
         self.substream = zmqstream.ZMQStream(self.sub_socket, self.loop)
         self.substream.on_recv(self.handle_recv)
         self.client = CircusClient(context=self.ctx, endpoint=endpoint)
         self.cmds = get_commands()
         self._pids = defaultdict(list)
         self._callbacks = dict()
-        self.collector = StatsCollector()
         self.publisher = StatsPublisher(stats_endpoint, self.ctx)
         self.running = False  # should the streamer be running?
         self.stopped = False  # did the collect started yet?
         self.circus_pids = {}
-
-    def publish_stats(self, watcher=None):
-        """Get and publish the stats for the given watcher"""
-        logger.debug('Publishing stats about {0}'.format(watcher))
-        process_name = None
-        for watcher, pid, stats in self.collector.collect_stats(
-                watcher, self.get_pids(watcher)):
-            if watcher == 'circus':
-                if pid in self.circus_pids:
-                    process_name = self.circus_pids[pid]
-
-            self.publisher.publish(watcher, process_name, pid, stats)
+        self.sockets = []
 
     def get_watchers(self):
         return self._pids.keys()
+
+    def get_sockets(self):
+        return self.sockets
 
     def get_pids(self, watcher=None):
         if watcher is not None:
@@ -59,26 +51,65 @@ class StatsStreamer(object):
         return chain(self._pid.values())
 
     def get_circus_pids(self):
-        # getting the circusd pid
+        # getting the circusd, circusd-stats and circushttpd pids
         res = self.client.send_message('dstats')
-        return {os.getpid(): 'circusd-stats',
+        pids = {os.getpid(): 'circusd-stats',
                 res['info']['pid']: 'circusd'}
 
+        httpd_pids = self.client.send_message('list', name='circushttpd')
+
+        if 'pids' in httpd_pids:
+            httpd_pids = httpd_pids['pids']
+            if len(httpd_pids) == 1:
+                pids[httpd_pids[0]] = 'circushttpd'
+        return pids
+
+    def _add_callback(self, name, start=True, kind='watcher'):
+        logger.debug('Callback added for %s' % name)
+
+        if kind == 'watcher':
+            klass = WatcherStatsCollector
+        elif kind == 'socket':
+            klass = SocketStatsCollector
+        else:
+            raise ValueError('Unknown callback kind %r' % kind)
+
+        self._callbacks[name] = klass(self, name, self.delay, self.loop)
+        if start:
+            self._callbacks[name].start()
+
     def _init(self):
-        self.circus_pids = self.get_circus_pids()
-        if 'circus' not in self._callbacks:
-            self._callbacks['circus'] = ioloop.PeriodicCallback(
-                    lambda: self.publish_stats("circus"),
-                    self.delay * 1000, self.loop)
-        self._callbacks['circus'].start()
         self._pids.clear()
+
         # getting the initial list of watchers/pids
         res = self.client.send_message('list')
 
         for watcher in res['watchers']:
+            if watcher in ('circusd', 'circushttpd', 'circusd-stats'):
+                # this is dealt by the special 'circus' collector
+                continue
+
             pids = self.client.send_message('list', name=watcher)['pids']
             for pid in pids:
                 self.append_pid(watcher, pid)
+
+        # getting the circus pids
+        self.circus_pids = self.get_circus_pids()
+        if 'circus' not in self._callbacks:
+            self._add_callback('circus')
+        else:
+            self._callbacks['circus'].start()
+
+        # getting the initial list of sockets
+        res = self.client.send_message('listsockets')
+        for sock in res['sockets']:
+            fd = sock['fd']
+            address = '%s:%s' % (sock['host'], sock['port'])
+            # XXX type / family ?
+            sock = socket.fromfd(fd, socket.AF_INET, socket.SOCK_STREAM)
+            self.sockets.append((sock, address, fd))
+
+        self._add_callback('sockets', kind='socket')
 
     def remove_pid(self, watcher, pid):
         if pid in self._pids[watcher]:
@@ -91,13 +122,12 @@ class StatsStreamer(object):
 
     def append_pid(self, watcher, pid):
         if watcher not in self._pids or len(self._pids[watcher]) == 0:
-            if watcher not in self._callbacks:
-                self._callbacks[watcher] = ioloop.PeriodicCallback(
-                        lambda: self.publish_stats(watcher),
-                        self.delay * 1000, self.loop)
             logger.debug('Starting the periodic callback for {0}'\
                          .format(watcher))
-            self._callbacks[watcher].start()
+            if watcher not in self._callbacks:
+                self._add_callback(watcher)
+            else:
+                self._callbacks[watcher].start()
 
         if pid in self._pids[watcher]:
             return