+++ circus/web/circushttpd.py
@@ -1,6 +1,7 @@
 import argparse
 import os
 import sys
+import socket
 
 try:
     from beaker.middleware import SessionMiddleware
@@ -12,13 +13,16 @@ try:
     from socketio.namespace import BaseNamespace
 
 except ImportError, e:
+    reqs = os.path.join(os.path.abspath(os.path.dirname(__file__)),
+                        'web-requirements.txt')
     raise ImportError('You need to install dependencies to run the webui. '\
                     + 'You can do so by using "pip install -r '
-                    + 'web-requirements.txt"\nInitial error: %s' % str(e))
+                    + '%s"\nInitial error: %s' % (reqs, str(e)))
 
 from circus.web.controller import LiveClient, CallError
 from circus.stats.client import StatsClient
-from circus import __version__
+from circus import __version__, logger
+from circus.util import configure_logger, LOG_LEVELS
 
 
 _DIR = os.path.dirname(__file__)
@@ -48,7 +52,7 @@ def route(*args, **kwargs):
                 global client
                 if client is None:
                     session = get_session()
-                    if 'endpoint' in session:
+                    if session.get('endpoint', None) is not None:
                         client = connect_to_endpoint(session['endpoint'])
                     else:
                         return redirect('/connect')
@@ -113,22 +117,40 @@ def watcher(name):
     return render_template('watcher.html', name=name)
 
 
+@route('/sockets', method='GET')
+def sockets():
+    return render_template('sockets.html')
+
+
 @route('/connect', method=['POST', 'GET'], ensure_client=False)
 def connect():
-    if request.method == 'GET':
+    """Connects to the stats client, using the endpoint that's passed in the
+    POST body.
+    """
+    def _ask_connection():
         return render_template('connect.html')
 
+    if request.method == 'GET':
+        return _ask_connection()
+
     elif request.method == 'POST':
+        # if we got an endpoint in the POST body, store it.
+        if request.forms.endpoint is None:
+            return _ask_connection()
+
+        endpoint = request.forms.endpoint
+
+        tmp_client = connect_to_endpoint(endpoint)
+        if not tmp_client.connected:
+            set_message('Impossible to connect')
+
         session = get_session()
-        session['endpoint'] = request.forms.endpoint
+        session['endpoint'] = endpoint
         session.save()
 
-        client = connect_to_endpoint(request.forms.endpoint)
-        if client.connected:
-            message = 'You are now connected'
-        else:
-            message = 'Impossible to connect'
-        set_message(message)
+        global client
+        client = tmp_client
+
         redirect('/')
 
 
@@ -152,6 +174,11 @@ def socketio(someid, socket_id):
 
 
 class StatsNamespace(BaseNamespace, RoomsMixin, BroadcastMixin):
+
+    def __init__(self, *args, **kwargs):
+        super(StatsNamespace, self).__init__(*args, **kwargs)
+        self._running = True
+
     def on_get_stats(self, msg):
         """This method is the one way to start a conversation with the socket.
         When sending a message here, the parameters are packt into the msg
@@ -171,6 +198,9 @@ class StatsNamespace(BaseNamespace, RoomsMixin, BroadcastMixin):
             - "stats-<watchername>-<pid>" sends the information about
               specific pids for the different watchers (works only if
               "get_processes" is set to True when calling this method)
+            - "socket-stats" send the aggregation information about
+               sockets.
+            - "socket-stats-<fd>" sends information about a particular fd.
         """
 
         # unpack the params
@@ -178,25 +208,52 @@ class StatsNamespace(BaseNamespace, RoomsMixin, BroadcastMixin):
         streamsWithPids = msg.get('watchersWithPids', [])
 
         # if we want to supervise the processes of a watcher, then send the
-        # list of pids trough a socket.
+        # list of pids trough a socket. If we asked about sockets, do the same
+        # with their fds
         for watcher in streamsWithPids:
-            pids = [int(pid) for pid in client.get_pids(watcher)]
-            channel = 'stats-{watcher}-pids'.format(watcher=watcher)
-            self.send_data(channel, pids=pids)
+            if watcher == "sockets":
+                fds = [s['fd'] for s in client.get_sockets()]
+                self.send_data('socket-stats-fds', fds=fds)
+            else:
+                pids = [int(pid) for pid in client.get_pids(watcher)]
+                channel = 'stats-{watcher}-pids'.format(watcher=watcher)
+                self.send_data(channel, pids=pids)
 
         # Get the channels that are interesting to us and send back information
         # there when we got them.
         stats = StatsClient(endpoint=client.stats_endpoint)
         for watcher, pid, stat in stats:
-            if watcher in streams or watcher in streamsWithPids:
-                if pid is None:  # means that it's the aggregation
-                    self.send_data('stats-{watcher}'.format(watcher=watcher),
-                                   mem=stat['mem'], cpu=stat['cpu'])
-                else:
-                    if watcher in streamsWithPids:
-                        self.send_data('stats-{watcher}-{pid}'\
-                                       .format(watcher=watcher, pid=pid),
-                                       mem=stat['mem'], cpu=stat['cpu'])
+            if self._running == False:
+                return
+
+            if watcher == 'sockets':
+                # if we get information about sockets and we explicitely
+                # requested them, send back the information.
+                if 'sockets' in streamsWithPids and 'fd' in stat:
+                    self.send_data('socket-stats-{fd}'.format(fd=stat['fd']),
+                                   **stat)
+                elif 'sockets' in streams and 'addresses' in stat:
+                    self.send_data('socket-stats', reads=stat['reads'],
+                                   adresses=stat['addresses'])
+            else:
+                available_watchers = streams + streamsWithPids + ['circus']
+                # these are not sockets but normal watchers
+                if watcher in available_watchers:
+                    if (watcher == 'circus'
+                            and stat.get('name', None) in available_watchers):
+                        self.send_data(
+                                'stats-{watcher}'.format(watcher=stat['name']),
+                                mem=stat['mem'], cpu=stat['cpu'])
+                    else:
+                        if pid is None:  # means that it's the aggregation
+                            self.send_data(
+                                    'stats-{watcher}'.format(watcher=watcher),
+                                    mem=stat['mem'], cpu=stat['cpu'])
+                        else:
+                            if watcher in streamsWithPids:
+                                self.send_data('stats-{watcher}-{pid}'\
+                                           .format(watcher=watcher, pid=pid),
+                                           mem=stat['mem'], cpu=stat['cpu'])
 
     def send_data(self, topic, **kwargs):
         """Send the given dict encoded into json to the listening socket on the
@@ -209,10 +266,22 @@ class StatsNamespace(BaseNamespace, RoomsMixin, BroadcastMixin):
                    endpoint=self.ns_name)
         self.socket.send_packet(pkt)
 
+    def recv_disconnect(self):
+        """When we receive a disconnect from the client, we want to make sure
+        that we close the socket we just opened at the begining of the stat
+        exchange."""
+        self._running = False
+
 
 # Utils
 
 class SocketIOServer(ServerAdapter):
+    def __init__(self, host='127.0.0.1', port=8080, **config):
+        super(SocketIOServer, self).__init__(host, port, **config)
+        self.fd = config.get('fd')
+        if self.fd is not None:
+            self.fd = int(self.fd)
+
     def run(self, handler):
         try:
             from socketio.server import SocketIOServer
@@ -226,7 +295,13 @@ class SocketIOServer(ServerAdapter):
 
         namespace = self.options.get('namespace', 'socket.io')
         policy_server = self.options.get('policy_server', False)
-        socket_server = SocketIOServer((self.host, self.port), handler,
+
+        if self.fd is not None:
+            sock = socket.fromfd(self.fd, socket.AF_INET, socket.SOCK_STREAM)
+        else:
+            sock = (self.host, self.port)
+
+        socket_server = SocketIOServer(sock, handler,
                                        namespace=namespace,
                                        policy_server=policy_server)
         handler.socket_server = socket_server
@@ -256,7 +331,10 @@ def run_command(func, message, redirect_url, redirect_on_error=None,
     kwargs = kwargs or {}
 
     try:
+        logger.debug('Running %r' % func)
         res = func(*args, **kwargs)
+        logger.debug('Result : %r' % res)
+
         if res['status'] != 'ok':
             message = "An error happened: %s" % res['reason']
     except CallError, e:
@@ -295,16 +373,23 @@ def render_template(template, **data):
 
 def main():
     parser = argparse.ArgumentParser(description='Run the Web Console')
+
+    parser.add_argument('--fd', help='FD', default=None)
     parser.add_argument('--host', help='Host', default='0.0.0.0')
     parser.add_argument('--port', help='port', default=8080)
     parser.add_argument('--server', help='web server to use',
                         default=SocketIOServer)
+    parser.add_argument('--endpoint', default=None,
+        help='Circus Endpoint. If not specified, Circus will ask you which '
+             'system you want to connect to')
     parser.add_argument('--version', action='store_true',
                      default=False, help='Displays Circus version and exits.')
-
-    args = parser.parse_args()
-    old = list(sys.argv)
-    sys.argv = []
+    parser.add_argument('--log-level', dest='loglevel', default='info',
+            choices=LOG_LEVELS.keys() + [key.upper() for key in
+                LOG_LEVELS.keys()],
+            help="log level")
+    parser.add_argument('--log-output', dest='logoutput', default='-',
+            help="log output")
 
     args = parser.parse_args()
 
@@ -312,10 +397,14 @@ def main():
         print(__version__)
         sys.exit(0)
 
-    try:
-        run(app, host=args.host, port=args.port, server=args.server)
-    finally:
-        sys.argv = old
+    # configure the logger
+    configure_logger(logger, args.loglevel, args.logoutput)
+
+    if args.endpoint is not None:
+        global client
+        client = connect_to_endpoint(args.endpoint)
+
+    run(app, host=args.host, port=args.port, server=args.server, fd=args.fd)
 
 
 if __name__ == '__main__':